Update script.js
Updated the code to enhance the emoji explosion feature by adding the uniqueness option to the emojisplosion function. This ensures that the location of the explosion is not repeated too frequently in the explosion. Additionally, increased the number of times the function runs to create more explosions. This change improves the visual effect of the feature."